+++ template-examples/spring-axon-gamerental/src/main/kotlin/example/gamerental/Install.kt
@@ -1,0 +1,10 @@
+package example.gamerental
+
+import example.gamerental.GameRentalTemplate
+import sift.core.template.save
+
+// project pom is configured to run this during maven's `install` phase:
+// saves "petclinic" template to ~/.local/share/sift/templates/petclinic.json
+fun main() {
+    GameRentalTemplate().save()
+}